Add files via upload
In this project we have Orange telecom's churn dataset. Orange S.A. formerly France Telecom S.A., is a french multinantional telecommunications corporation. This dataset consists of a cleaned customer activity data, along with a churn label specifying wheather a customer cancelled the subscription. Our task was to analyze the data and give the recommendations to deal with the problem.
 At first we checked the data weather it is properly cleaned or not, then started exploratory data analysis in which we performed univariate analysis then bivariate analysis and then multivariate analysis and found out that there are some states where churned rate is very high i.e. around 25% of customers in those states are churning then we also found out that the major reson behind churning is highe service centre calls and having international plan has also significant impact on churning.